server/requirements.txt
bcrypt==3.2.0
Flask==1.1.2
python-jwt==3.3.0
mysql-connector-python==8.0.22
Flask-Cors==1.10.3
Werkzeug==1.0.1
requests==2.25.1